pyproject.toml
[tool.poetry]
name = "proxy"
version = "0.1.0"
description = ""
authors = ["tclesius <tomclesius@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp-ip-rotator = "^1.0"
fastapi = "^0.105.0"
pyyaml = "^6.0.1"
tqdm = "^4.66.1"
uvicorn = "^0.24.0.post1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
